requirements.txt
accelerate==0.19.0
diffusers @ git+https://github.com/huggingface/diffusers@1994dbcb5e62bd8d0c60e5d5d6bf4b580653c74c
huggingface-hub==0.15.1
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
PyYAML==6.0
requests==2.31.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.2
urllib3==1.26.6
imageio==2.31.3
scipy==1.10.1
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117